add a LineBasedModule for the common task of parsing line-by-line input
