+++ csdr/module/msk144.py
@@ -1,6 +1,6 @@
 from pycsdr.types import Format
 from pycsdr.modules import ExecModule
-from csdr.module import ThreadModule
+from csdr.module import LineBasedModule
 from owrx.wsjt import WsjtParser, Msk144Profile
 import pickle
 
@@ -17,40 +17,17 @@ class Msk144Module(ExecModule):
         )
 
 
-class ParserAdapter(ThreadModule):
+class ParserAdapter(LineBasedModule):
     def __init__(self):
-        self.retained = bytes()
         self.parser = WsjtParser()
         self.dialFrequency = 0
+        self.profile = Msk144Profile()
         super().__init__()
 
-    def run(self):
-        profile = Msk144Profile()
-
-        while self.doRun:
-            data = self.reader.read()
-            if data is None:
-                self.doRun = False
-            else:
-                self.retained += data
-                lines = self.retained.split(b"\n")
-
-                # keep the last line
-                # this should either be empty if the last char was \n
-                # or an incomplete line if the read returned early
-                self.retained = lines[-1]
-
-                # parse all completed lines
-                for line in lines[0:-1]:
-                    # actual messages from msk144decoder should start with "*** "
-                    if line[0:4] == b"*** ":
-                        self.writer.write(pickle.dumps(self.parser.parse(profile, self.dialFrequency, line[4:])))
-
-    def getInputFormat(self) -> Format:
-        return Format.CHAR
-
-    def getOutputFormat(self) -> Format:
-        return Format.CHAR
+    def process(self, line: bytes):
+        # actual messages from msk144decoder should start with "*** "
+        if line[0:4] == b"*** ":
+            return self.parser.parse(self.profile, self.dialFrequency, line[4:])
 
     def setDialFrequency(self, frequency: int) -> None:
         self.dialFrequency = frequency
