+++ owrx/adsb/dump1090.py
@@ -1,9 +1,8 @@
 from pycsdr.modules import ExecModule, Writer, TcpSource
 from pycsdr.types import Format
-from csdr.module import LogWriter, ThreadModule
+from csdr.module import LogWriter, LineBasedModule
 from owrx.socket import getAvailablePort
 import time
-import pickle
 
 import logging
 
@@ -51,38 +50,8 @@ class Dump1090Module(ExecModule):
             self.tcpSource.setWriter(writer)
 
 
-class RawDeframer(ThreadModule):
-    def __init__(self):
-        self.retained = bytes()
-        super().__init__()
-
-    def getInputFormat(self) -> Format:
-        return Format.CHAR
-
-    def getOutputFormat(self) -> Format:
-        return Format.CHAR
-
-    def run(self):
-        while self.doRun:
-            data = self.reader.read()
-            if data is None:
-                self.doRun = False
-            else:
-                self.retained += data
-                lines = self.retained.split(b"\n")
-
-                # keep the last line
-                # this should either be empty if the last char was \n
-                # or an incomplete line if the read returned early
-                self.retained = lines[-1]
-
-                # log all completed lines
-                for line in lines[0:-1]:
-                    parsed = self.parse(line)
-                    if parsed is not None:
-                        self.writer.write(pickle.dumps(parsed))
-
-    def parse(self, line):
+class RawDeframer(LineBasedModule):
+    def process(self, line: bytes):
         if line.startswith(b'*') and line.endswith(b';') and len(line) in [16, 30]:
             return bytes.fromhex(line[1:-1].decode())
         else:
